Change int to long.

+++ src/main/java/com/github/thomaskeschl/eu/core/NaturalNumber.java
@@ -4,9 +4,9 @@ package com.github.thomaskeschl.eu.core;
  * Created by tkeschl on 2/6/2015.
  */
 public class NaturalNumber {
-    private final int num;
+    private final long num;
 
-    public NaturalNumber(int num) {
+    public NaturalNumber(long num) {
         if(num < 0) {
             String message = String.format("Natural numbers cannot be less than 0! Value: %1$s", num);
             throw new IllegalArgumentException(message);
